chore: Resolved a few errors.

+++ src/Bot/Engine/Value/Value.h
@@ -10,7 +10,6 @@
 
 #include "AiObject.h"
 #include "ObjectGuid.h"
-#include "PerfMonitor.h"
 #include "Timer.h"
 #include "Unit.h"
 
@@ -24,7 +23,11 @@ public:
     float radius;
     float angle;
     uint32 timestamp;
-    int GetAngleRangeIndex() { return (angle + 2 * M_PI) / (M_PI / 2); }  // [0, 7)
+    int GetAngleRangeIndex()
+    {
+        // [0, 7)
+        return int((angle + 2 * M_PI) / (M_PI / 2));
+    }
 };
 
 struct CreatureData;
@@ -83,7 +86,7 @@ public:
             time_t now = getMSTime();
             if (!lastCheckTime || now - lastCheckTime >= checkInterval)
             {
-                lastCheckTime = now;
+                lastCheckTime = uint32_t(now);
                 // PerfMonitorOperation* pmo = sPerfMonitor.start(PERF_MON_VALUE, this->getName(),
                 // this->context ? &this->context->performanceStack : nullptr);
                 value = Calculate();
@@ -116,7 +119,7 @@ public:
             time_t now = getMSTime();
             if (!lastCheckTime || now - lastCheckTime >= checkInterval)
             {
-                lastCheckTime = now;
+                lastCheckTime = uint32_t(now);
                 // PerfMonitorOperation* pmo = sPerfMonitor.start(PERF_MON_VALUE, this->getName(),
                 // this->context ? &this->context->performanceStack : nullptr);
                 value = Calculate();
@@ -273,7 +276,7 @@ class Uint32CalculatedValue : public CalculatedValue<uint32>
 {
 public:
     Uint32CalculatedValue(PlayerbotAI* botAI, std::string const name = "value", int checkInterval = 1)
-        : CalculatedValue<uint32>(botAI, name, checkInterval)
+        : CalculatedValue<uint32>(botAI, name, uint32_t(checkInterval))
     {
     }
 
@@ -284,7 +287,7 @@ class FloatCalculatedValue : public CalculatedValue<float>
 {
 public:
     FloatCalculatedValue(PlayerbotAI* botAI, std::string const name = "value", int checkInterval = 1)
-        : CalculatedValue<float>(botAI, name, checkInterval)
+        : CalculatedValue<float>(botAI, name, uint32_t(checkInterval))
     {
     }
 
@@ -295,7 +298,7 @@ class BoolCalculatedValue : public CalculatedValue<bool>
 {
 public:
     BoolCalculatedValue(PlayerbotAI* botAI, std::string const name = "value", int checkInterval = 1)
-        : CalculatedValue<bool>(botAI, name, checkInterval)
+        : CalculatedValue<bool>(botAI, name, uint32_t(checkInterval))
     {
     }
